src/lib.rs:refactor
